Alterando a remoção e adição na árvore usando exceção Ainda não está pronto, foi removido o código de gravar no arquivo, ainda será reposto.

+++ sistema-eleitoral/src/pi/arvore/Arvore.java
@@ -19,7 +19,7 @@ public interface Arvore <T extends Elemento>{
 	 * @return removido
 	 * @param elemento
 	 */
-	public boolean remove(Integer elemento);
+	public boolean remove(long elemento);
 	
 	/**
 	 * Retorna a raiz da árvore
